fix: don't throw when parsing the tsconfig.json
Apparently TypeScript will throw an error of the tsconfig.json matches
no files. This is reasonable for regular uses of TypeScript but
unfortunately in our case it can be that the Framer X project is empty
before importing has taken place.

If the canvas.tsx was not yet created, an error would be thrown when
parsing the tsconfig.json, this change instead makes the error visible
but continues with the process.

Read more: https://stackoverflow.com/questions/41211566/tsconfig-json-buildno-inputs-were-found-in-config-file